simplify tests
Since the tests are using supertest there is no need to "start" the app
or even call .listen() on it in order to send requests to it.

Removing the explicit listen both simplifies the tests as well as
ensures that they use an ephemeral port, which makes it safer to run the
tests on CI where multiple tests may be running at the same time and
should not use fixed ports because they may conflict.