disable sending of email from CI
There is no reason to unconditionally send an email using invalid
credentials, every time the app is loaded. If CI doesn't finish
immediately, the app is loaded for long enough that it actually tries to
send the email and then crashes when it fails.

It appears this was introduced in #12, but the tests weren't being run
on Mac at the time, so the problem went unnoticed until they were.